pyproject.toml
[project]
name = "pycomicconverter"
version = "0.1.9"
description = "Utility to convert comic files"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.3.0",
    "pymupdf>=1.26.3",
]

[project.optional-dependencies]
build = ["pyinstaller>=6.0"]
